Improved Controls
Replaced Tutorial Controls with Click and Drag Controls
Added Camera Reset (Rotation & Position) on Space
Set Scale to 1 instead of 0.01
Camera now starts at center of map
Converted Colours to floats
Crappy zoom added using scroll wheel to modify FOV

+++ 3DRenderer/Island_Utils.h
@@ -9,7 +9,7 @@ namespace Island_Utils
 {
 	vector<vector<int>> FlattenWater(vector<vector<int>> *elevationMap);
 	vector<vector<int>> CalculateBiomes(vector<vector<int>> *elevationMap, vector<vector<int>> *moistureMap);
-	vector<int> GetBiomeColour(int biome);
+	vector<float> GetBiomeColour(int biome);
 	vector<vector<int>> MakeCircularGradient(int width, int height, int maxValue = 255, int minValue = 0);
 	vector<vector<int>> InterpolateBitmaps(vector<vector<int>> *bmp1, vector<vector<int>> *bmp2, double bmp1Coeff = 0.5,
 		double bmp2Coeff = 0.5, int offset = 0, int maxValue = 255, int minValue = 0);
